新建文件夹/files/legal_rag_v2/legal_rag_v2/requirements.txt
langchain>=1.2.0
langchain-community>=0.4.0
langchain-core>=1.2.0
langchain-openai>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=1.5.0
rank-bm25>=0.2.2
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
mineru[full]
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
